// apps/pwa/src/components/pages/nestedRoutes/old-signs/SignsContainer/SignsLists/BloodPressureListContainer/type.d.ts
import { BloodPressureType } from '@components/pages/mainRoutes/calendar/__hooks__/useCalendarGetData/types';

export type BloodPressureListPropsType = {
  bloodPressureList: BloodPressureType[];
};

type BloodPressureState = {
  text: string;
  border: string;
  result: string;
};

export type InitailBloodPressureContextType = {
  normal: BloodPressureState;
  danger: BloodPressureState;
};

export type CheckBloodPressureStatusType = (high: number, low: number) => keyof InitailBloodPressureContextType;